split views from main
